Move logic to a separate folder, and create useResponsiveValues hook
