+++ react/useResponsiveValues.ts
@@ -1,0 +1,3 @@
+import { useResponsiveValues } from './hooks/useResponsiveValue'
+
+export default useResponsiveValues
